made the utils.chronological module part of the utils.timestamps module instead and merged testing files.
