+++ tests/utils/test_timestamps.py
@@ -1,8 +1,11 @@
-"""Provides Unit Tests for the `abis_mapping.utils.coords` module"""
+"""Provides Unit Tests for the `abis_mapping.utils.timestamps` module"""
 
 
 # Third-Party
 import pytest
+
+# Standard
+import datetime
 
 # Local
 from abis_mapping import utils
@@ -64,3 +67,24 @@ def test_timestamp_parse_invalid(raw: Any) -> None:
     # Parse Invalid Timestamps
     with pytest.raises(ValueError):
         utils.timestamps.parse_timestamp(raw)
+
+
+def test_is_chronologically_ordered() -> None:
+    """Tests the is_chronologically_ordered() function."""
+
+    # Define scenario lists
+    ordered_datetimes = [
+        datetime.datetime(2022, 9, 11, 15, 15, 15),
+        datetime.datetime(2023, 9, 11, 15, 15, 15),
+        datetime.datetime(2023, 10, 11, 15, 15, 15),
+    ]
+    unordered_datetimes = [
+        datetime.datetime(2022, 9, 11, 15, 15, 15),
+        datetime.datetime(2023, 10, 11, 15, 15, 15),
+        datetime.datetime(2023, 9, 11, 15, 15, 15),
+    ]
+
+    # Check chronologically increasing
+    assert utils.timestamps.is_chronologically_ordered(ordered_datetimes)
+    assert not utils.timestamps.is_chronologically_ordered(unordered_datetimes)
+
